+++ packages/cli/generation/ir-migrations/src/migrations/v67-to-v66/__test__/migrateFromV67ToV66.test.ts
@@ -267,6 +267,7 @@ describe("migrateFromV67ToV66", () => {
                             valueType: IrVersions.V67.TypeReference.primitive({ v1: "STRING", v2: undefined }),
                             propertyAccess: undefined,
                             defaultValue: undefined,
+                            xml: undefined,
                             v2Examples: undefined
                         }
                     ]